Moved the Usecases to kommunicate module
