+++ kommunicate/src/main/java/io/kommunicate/usecase/UseCase.kt
@@ -1,4 +1,4 @@
-package com.applozic.mobicomkit.usecase
+package io.kommunicate.usecase
 
 interface UseCase<T> {
     suspend fun execute(): T
